# Cache template lookups in Path