--- a/lib/dry/view/path.rb
+++ b/lib/dry/view/path.rb
@@ -7,6 +7,16 @@ module Dry
 
       attr_reader :dir, :root
 
+      def self.cached_templates
+        @cached_templates ||= {}
+      end
+
+      def self.fetch_from_cache_or_store(key)
+        cached_templates.fetch(key) do
+          cached_templates[key] = yield
+        end
+      end
+
       def initialize(dir, options = {})
         @dir = Pathname(dir)
         @root = Pathname(options.fetch(:root, dir))
@@ -32,8 +42,10 @@ module Dry
 
       # Search for a template using a wildcard for the engine extension
       def template?(name, format)
-        glob = dir.join("#{name}.#{format}.*")
-        Dir[glob].first
+        unique_key = "#{dir}/#{name}.#{format}"
+        self.class.fetch_from_cache_or_store(unique_key) do
+          Dir[dir.join("#{name}.#{format}.*")].first
+        end
       end
     end
   end