Fix #94 

Thanks to @solnic for pointing out that the method `Path#template?` is very expensive. I decided to add some simple cached for storing the result of looking for template.

The performance results are positive after adding this functionality.

Before:

```bash
Warming up --------------------------------------
   action_controller     1.000  i/100ms
            dry-view     1.000  i/100ms
Calculating -------------------------------------
   action_controller      6.552  (± 0.0%) i/s -     33.000  in   5.038370s
            dry-view      1.289  (± 0.0%) i/s -      7.000  in   5.430740s

Comparison:
   action_controller:        6.6 i/s
            dry-view:        1.3 i/s - 5.08x  slower
```

After 

```bash
Warming up --------------------------------------
   action_controller     1.000  i/100ms
            dry-view     1.000  i/100ms
Calculating -------------------------------------
   action_controller      6.816  (± 0.0%) i/s -     35.000  in   5.136691s
            dry-view      3.750  (± 0.0%) i/s -     19.000  in   5.068494s

Comparison:
   action_controller:        6.8 i/s
            dry-view:        3.8 i/s - 1.82x  slower
```