@timriley 2018-12-21
This is part of #97 now